+++ tests/Transformers/ImageVectorizerTest.php
@@ -3,6 +3,7 @@
 namespace Rubix\ML\Tests\Transformers;
 
 use Rubix\ML\Datasets\Unlabeled;
+use Rubix\ML\Transformers\Stateful;
 use Rubix\ML\Transformers\Transformer;
 use Rubix\ML\Transformers\ImageResizer;
 use Rubix\ML\Transformers\ImageVectorizer;
@@ -31,10 +32,10 @@ class ImageVectorizerTest extends TestCase
     protected function setUp() : void
     {
         $this->dataset = Unlabeled::quick([
-            [imagecreatefromjpeg('tests/space.jpg')],
+            [imagecreatefromjpeg('tests/space.jpg'), 'something else'],
         ]);
 
-        $this->transformer = new ImageVectorizer(3);
+        $this->transformer = new ImageVectorizer(false);
     }
     
     /**
@@ -44,19 +45,20 @@ class ImageVectorizerTest extends TestCase
     {
         $this->assertInstanceOf(ImageVectorizer::class, $this->transformer);
         $this->assertInstanceOf(Transformer::class, $this->transformer);
+        $this->assertInstanceOf(Stateful::class, $this->transformer);
     }
     
     /**
      * @test
      */
-    public function transform() : void
+    public function fitTransform() : void
     {
         $this->dataset->apply(new ImageResizer(3, 3));
 
         $this->dataset->apply($this->transformer);
     
         $outcome = [
-            [60, 35, 22, 102, 66, 53, 73, 44, 29, 79, 49, 36, 89, 57, 45, 56,
+            ['something else', 60, 35, 22, 102, 66, 53, 73, 44, 29, 79, 49, 36, 89, 57, 45, 56,
                 32, 21, 85, 53, 44, 75, 49, 43, 34, 18, 12],
         ];