+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     - Changed order of Missing Data Imputer params
     - Renamed high-level resource type to image
     - Added comb (n choose k) as a global function
+    - Image Vectorizer now has grayscale option
 
 - 0.0.18-beta
     - Now requires PHP 7.2 and above